use css and update naming for clarity for add new competition box
