+++ src/components/AddCompetitionButton/AddCompetitionButton.jsx
@@ -1,0 +1,27 @@
+import { modals } from "@mantine/modals";
+import { Text, Button } from "@mantine/core";
+import { IconCirclePlus } from "@tabler/icons-react";
+import AddCompetitionModal from "./AddCompetitionModal.jsx";
+import styles from "./AddCompetitionButton.module.css";
+
+function AddCompetitionBox() {
+  return (
+    <Button
+      variant="filled"
+      size="xl"
+      leftSection={<IconCirclePlus size={27} />}
+      className={styles.button}
+      onClick={() => {
+        modals.open({
+          title: <Text class="modalHeading">New Competition</Text>,
+          children: <AddCompetitionModal />,
+          centered: true,
+        });
+      }}
+    >
+      Add new competition
+    </Button>
+  );
+}
+
+export default AddCompetitionBox;
