+++ src/components/AddCompetitionButton/AddCompetitionButton.module.css
@@ -1,0 +1,6 @@
+.button {
+  width: 80%;
+  height: 50%;
+  border-radius: 1.5rem;
+  background-color: #92809a;
+}
