+++ src/components/AddCompetitionButton/AddCompetitionModal.jsx
